+++ Source/RTEngine/PhongMaterial.cpp
@@ -3,19 +3,19 @@
 
 using namespace RealRT;
 
-PhongMaterial::PhongMaterial(const Vector3D &color, float reflectance, float refractance, float indexOfRefraction, float diffuseScalar, float specularScalar)
+PhongMaterial::PhongMaterial(const Vector3D &color, double reflectance, double refractance, double indexOfRefraction, double diffuseScalar, double specularScalar)
     : Material(color, reflectance, refractance, indexOfRefraction, diffuseScalar, specularScalar)
 {
 
 }
 
-float PhongMaterial::BidirectionReflectanceDistributionFunction(const Vector3D &in, const Vector3D &out, const Vector3D &norm)  const
+double PhongMaterial::BidirectionReflectanceDistributionFunction(const Vector3D &in, const Vector3D &out, const Vector3D &norm)  const
 {
-    float ratio = 0.f;
+    double ratio = 0.0;
 
     //take the dot product of the unit normal against the vector to the light. Note that
     //since both are unit vectors, this value is the cosine of the angle between them
-    float dot = norm * out;
+    double dot = norm * out;
 
     //use only positive cosine values to shade only the parts facing the light source
     //
@@ -28,7 +28,7 @@ float PhongMaterial::BidirectionReflectanceDistributionFunction(const Vector3D &
     }
 
     Vector3D reflectionDir;
-    reflectionDir = out - 2.f * (out * norm) * norm;
+    reflectionDir = out - 2.0 * (out * norm) * norm;
     dot = reflectionDir * in;
 
     if(dot > 0)
