+++ Source/RTEngine/Vector3D.hpp
@@ -11,17 +11,17 @@ namespace RealRT
         public:
             Vector3D(void) = default;
 
-            Vector3D(float i, float j, float k);
+            Vector3D(double i, double j, double k);
 
-            explicit Vector3D(const std::array<float, 3> values);
+            explicit Vector3D(const std::array<double, 3> values);
 
-            Vector3D(std::initializer_list<float> list);
+            Vector3D(std::initializer_list<double> list);
 
-            float I(void) const;
+            double I(void) const;
 
-            float J(void) const;
+            double J(void) const;
 
-            float K(void) const;
+            double K(void) const;
 
             Vector3D Add(const Vector3D &v) const;
 
@@ -29,21 +29,21 @@ namespace RealRT
 
             Vector3D Normalize(void) const;
 
-            float Distance(const Vector3D &rhs) const;
+            double Distance(const Vector3D &rhs) const;
 
-            float Magnitude(void) const;
+            double Magnitude(void) const;
 
-            Vector3D Scale(float scalar) const;
+            Vector3D Scale(double scalar) const;
 
-            Vector3D Clip(float maxVal) const;
+            Vector3D Clip(double maxVal) const;
 
             Vector3D Weight(const Vector3D &dimWeights) const;
 
-            float Dot(const Vector3D &rhs) const;
+            double Dot(const Vector3D &rhs) const;
 
             Vector3D Cross(const Vector3D &rhs) const;
 
-            float operator [](int i) const;
+            double operator [](int i) const;
 
             bool operator ==(const Vector3D &rhs) const;
 
@@ -65,13 +65,13 @@ namespace RealRT
 
             Vector3D &operator -=(const Vector3D &rhs);
 
-            friend Vector3D operator *(float scalar, const Vector3D &v);
+            friend Vector3D operator *(double scalar, const Vector3D &v);
 
-            friend Vector3D operator *(const Vector3D &v, float scalar);
+            friend Vector3D operator *(const Vector3D &v, double scalar);
 
-            float operator *(const Vector3D &rhs) const;
+            double operator *(const Vector3D &rhs) const;
 
-            Vector3D &operator *=(float scalar);
+            Vector3D &operator *=(double scalar);
 
             Vector3D operator %(const Vector3D &rhs) const;
 
@@ -83,7 +83,7 @@ namespace RealRT
 
         private:
 
-            float _I = 0.f, _J = 0.f, _K = 0.f;
+            double _I = 0.0, _J = 0.0, _K = 0.0;
         };
 
 }
